fix(config-transfer): derive stored bandwidth fields
