+++ app/client/modules/auth/routes/download-recovery-key.tsx
@@ -1,12 +1,12 @@
 import { useMutation } from "@tanstack/react-query";
 import { useNavigate } from "@tanstack/react-router";
-import { AlertTriangle, Download } from "lucide-react";
+import { AlertTriangle, ArrowRight, CheckCircle2, Download, Upload } from "lucide-react";
 import { useState } from "react";
 import { toast } from "sonner";
 import { downloadResticPasswordMutation } from "~/client/api-client/@tanstack/react-query.gen";
 import { AuthLayout } from "~/client/components/auth-layout";
 import { Alert, AlertDescription, AlertTitle } from "~/client/components/ui/alert";
-import { Button } from "~/client/components/ui/button";
+import { Button, buttonVariants } from "~/client/components/ui/button";
 import { Input } from "~/client/components/ui/input";
 import { Label } from "~/client/components/ui/label";
 import { downloadFile } from "~/client/lib/download";
@@ -31,7 +31,11 @@ export function DownloadRecoveryKeyPage({ passwordAuthSupported, hasPassword, us
 	const navigate = useNavigate();
 	const [password, setPassword] = useState("");
 	const [blockedMessage, setBlockedMessage] = useState<string | null>(null);
+	const [importWarnings, setImportWarnings] = useState<string[] | null>(null);
 	const isDesktopRuntime = runtime === "desktop";
+	const goToVolumes = () => {
+		void navigate({ to: "/volumes", replace: true });
+	};
 
 	const downloadResticPassword = useMutation({
 		...downloadResticPasswordMutation(),
@@ -44,7 +48,7 @@ export function DownloadRecoveryKeyPage({ passwordAuthSupported, hasPassword, us
 
 			toast.success("Recovery key downloaded successfully!");
 			setBlockedMessage(null);
-			void navigate({ to: "/volumes", replace: true });
+			goToVolumes();
 		},
 		onError: (error) => {
 			const message = parseError(error)?.message;
@@ -72,12 +76,52 @@ export function DownloadRecoveryKeyPage({ passwordAuthSupported, hasPassword, us
 		if (!userId) return;
 
 		document.cookie = `${RECOVERY_KEY_DOWNLOAD_SKIPPED_COOKIE_NAME}=${userId}; path=/; max-age=${RECOVERY_KEY_DOWNLOAD_SKIPPED_COOKIE_MAX_AGE}`;
-		void navigate({ to: "/volumes", replace: true });
+		goToVolumes();
 	};
 
-	const handleImportSuccess = () => {
-		void navigate({ to: "/volumes", replace: true });
-	};
+	if (importWarnings !== null) {
+		const hasWarnings = importWarnings.length > 0;
+		const description = hasWarnings
+			? "Your configuration was imported, but some items need your attention."
+			: "Your configuration is ready to use.";
+		const alertVariant = hasWarnings ? "warning" : "default";
+		const alertTitle = hasWarnings ? "Review required" : "Import complete";
+		const continueLabel = hasWarnings ? "I understand, continue" : "Continue";
+
+		return (
+			<AuthLayout title="Configuration imported" description={description}>
+				<div className="space-y-6">
+					<Alert variant={alertVariant}>
+						{hasWarnings ? <AlertTriangle className="size-5" /> : <CheckCircle2 className="size-5" />}
+						<AlertTitle>{alertTitle}</AlertTitle>
+						<AlertDescription className="space-y-3">
+							{hasWarnings ? (
+								<>
+									<p>
+										The export contains settings and credentials, not backup data. Restore or mount
+										these local resources before using them:
+									</p>
+									<ul className="list-disc space-y-2 break-words pl-5">
+										{importWarnings.map((warning) => (
+											<li key={warning}>{warning}</li>
+										))}
+									</ul>
+									<p>Affected schedules stay disabled until you validate their paths.</p>
+								</>
+							) : (
+								<p>Your repositories, volumes, schedules, and notifications were imported.</p>
+							)}
+						</AlertDescription>
+					</Alert>
+
+					<Button variant="primary" className="w-full" onClick={goToVolumes}>
+						{continueLabel}
+						<ArrowRight className="ml-2 size-4" />
+					</Button>
+				</div>
+			</AuthLayout>
+		);
+	}
 
 	return (
 		<AuthLayout
@@ -153,7 +197,15 @@ export function DownloadRecoveryKeyPage({ passwordAuthSupported, hasPassword, us
 
 			<div className="my-6 border-t border-border/60" />
 
-			<ConfigImportForm onImportSuccess={handleImportSuccess} />
+			<details>
+				<summary className={buttonVariants({ variant: "link", className: "list-none px-0" })}>
+					<Upload className="mr-2 size-4" />
+					Import configuration
+				</summary>
+				<div className="mt-6">
+					<ConfigImportForm onSuccess={setImportWarnings} />
+				</div>
+			</details>
 		</AuthLayout>
 	);
 }
