+++ app/client/modules/repositories/components/create-repository-form.tsx
@@ -128,8 +128,9 @@ export const CreateRepositoryForm = ({
 
 	const { watch, setValue } = form;
 
-	const watchedBackend = watch("backend");
-	const watchedIsExistingRepository = watch("isExistingRepository");
+	const backend = watch("backend");
+	const isExisting = watch("isExistingRepository");
+	const exactPath = mode === "update" || isExisting === true;
 
 	const [passwordMode, setPasswordMode] = useState<"default" | "custom">("default");
 
@@ -268,34 +269,35 @@ export const CreateRepositoryForm = ({
 					)}
 				/>
 
-				<FormField
-					control={form.control}
-					name="isExistingRepository"
-					render={({ field }) => (
-						<FormItem className="flex flex-row items-center space-x-3">
-							<FormControl>
-								<Checkbox
-									checked={field.value}
-									disabled={mode === "update"}
-									onCheckedChange={(checked) => {
-										field.onChange(checked);
-										if (!checked) {
-											setPasswordMode("default");
-											setValue("customPassword", undefined);
-										}
-									}}
-								/>
-							</FormControl>
-							<div className="space-y-1">
-								<FormLabel>Import existing repository</FormLabel>
-								<FormDescription>
-									Check this if the repository already exists at the specified location
-								</FormDescription>
-							</div>
-						</FormItem>
-					)}
-				/>
-				{watchedIsExistingRepository && (
+				{mode === "create" && (
+					<FormField
+						control={form.control}
+						name="isExistingRepository"
+						render={({ field }) => (
+							<FormItem className="flex flex-row items-center space-x-3">
+								<FormControl>
+									<Checkbox
+										checked={field.value}
+										onCheckedChange={(checked) => {
+											field.onChange(checked);
+											if (!checked) {
+												setPasswordMode("default");
+												setValue("customPassword", undefined);
+											}
+										}}
+									/>
+								</FormControl>
+								<div className="space-y-1">
+									<FormLabel>Import existing repository</FormLabel>
+									<FormDescription>
+										Check this if the repository already exists at the specified location
+									</FormDescription>
+								</div>
+							</FormItem>
+						)}
+					/>
+				)}
+				{isExisting && (
 					<>
 						<FormItem>
 							<FormLabel>Repository Password</FormLabel>
@@ -350,16 +352,16 @@ export const CreateRepositoryForm = ({
 					</>
 				)}
 
-				{watchedBackend === "local" && <LocalRepositoryForm form={form} />}
-				{watchedBackend === "s3" && <S3RepositoryForm form={form} />}
-				{watchedBackend === "r2" && <R2RepositoryForm form={form} />}
-				{watchedBackend === "gcs" && <GCSRepositoryForm form={form} />}
-				{watchedBackend === "azure" && <AzureRepositoryForm form={form} />}
-				{watchedBackend === "rclone" && <RcloneRepositoryForm form={form} />}
-				{watchedBackend === "rest" && <RestRepositoryForm form={form} />}
-				{watchedBackend === "sftp" && <SftpRepositoryForm form={form} />}
+				{backend === "local" && <LocalRepositoryForm form={form} exactPath={exactPath} />}
+				{backend === "s3" && <S3RepositoryForm form={form} />}
+				{backend === "r2" && <R2RepositoryForm form={form} />}
+				{backend === "gcs" && <GCSRepositoryForm form={form} />}
+				{backend === "azure" && <AzureRepositoryForm form={form} />}
+				{backend === "rclone" && <RcloneRepositoryForm form={form} />}
+				{backend === "rest" && <RestRepositoryForm form={form} />}
+				{backend === "sftp" && <SftpRepositoryForm form={form} />}
 
-				{watchedBackend && watchedBackend !== "local" && <AdvancedForm form={form} />}
+				{backend && backend !== "local" && <AdvancedForm form={form} />}
 
 				{mode === "update" && (
 					<Button type="submit" className="w-full" loading={loading}>
