+++ app/server/modules/system/__fixtures__/config-transfer/v1-full.zbex
@@ -1,1 +1,1 @@
-zbcfg:v1:{"envelopeVersion":1,"kdf":{"name":"scrypt","salt":"c6CWEHfFkUF5_eMGCd7AIA","cost":65536,"blockSize":8,"parallelization":2,"keyBytes":32},"cipher":{"name":"aes-256-gcm","nonce":"I2DJagqb5x_0VNiT","authTag":"q7Rii5LAlvyTomx5W8_sHQ"},"ciphertext":"w3paS39HYv78EgH9KGZ6-n-C9GALESx51XUwab9cunnYt9GayGf_a8pvgipP89mPTJTLfQg9m58TZ2fSAvRmeUkiW5Zo_nFtt8xccU1eqmG4rZNGJAdwDq_YDaHi9LbW_n3bm8eHQyDMMeenpzYWNZXP84i3ToZ72V8FDsxeARNQ5FnngqShqK7k58XxyQ9YjLRHGqcE3n5lWqhd-2Bkewv1HuOyJUQcS96WUkouRVelPw7qxatG7UtZYNzONbh5EE6eQOfgLpHnGU_en7E3BRW155ql4eauTL8XF6HiY8pJvl82xbtDCLWPx4Ve-gCczLjjVsOO3ICVCsdBNlaGbrV0uMrBP_wD-ZxAYAUV8GemUTjqTJggEPkJz9D4yRDGu4aCiu-cboryCcQ78z3QpSutKZG_Dk0EnvwTKX-z4PqciJLCcVG5F0BXXsJfIIXeIwjuiNfQWZGoGoWetonH7JBtFoMITdNscieQtSYo8kM79D-EUv5sJjxmRSoBrOPkrTkb8pUznKh_paeqhNPPq_6yqsukGPQIcbSAj0tJ2yevM1qyOjIQN83EVxtnKLwH9Sjq5uhpKKSyl2uGxEE0qWCZ8ON51I3Tw7KFvfLfXy8cHqdsU_RKLTsF7q4JIQsC5DcM_SWBxNg77H-vhGVhYny-oiVOfO01jWr5hvf0nRUqVEU29aQXmHBvQQgNV9dG9MSRdtJVVHu_x6mQcLGmKW9ThgAwATyRtnjO8MC6wdwBqAudlHuP11OU3dfYjgTibz5mxqrERPRBoJbXngqQ4TrPFJUKejchFIS4S-wNbDjsEPihNJqlnxOYK5JImdsDB-mVUAFCQvECuABQXAww3KzgtyopJc-vVUfmn1PwrzZdRJCIbBN_UYsxN_AXKctBFhsTbqWHt3BKovemsE1qLWc0Bdg2NeYWrdvzc29jWAnESZuC8DbC5It2NTENqKCJs0Zlw1c776fcLRmH6AJu0SrUWWF-tefpuyYwwEH1Nl0NewHOWxGZJ2-JqtaqBngsKy55XAUsH8pOVoXIAasTL033Cie4YkJfh2WDDIio0odRMBFj8yLv-5xGJcU1qL7p0SlLUQSNmCDCjE_Q1MJQ1V_jORGoMeu5AYjwtj9V-S8WkoMMG7s9Vxpr56kaQS-LoWPPpSwOW201SytRqF5eMKVy7nzIFUq2ZbAVMTJH13d-9uWszaBM5skILYZgtl_x-PykXwMp7wQIbTVVHoXrJbCLBiuWmR5G16BliVI3Mzg7iNkJRDKXrlD2WqzC42x36b-TdN9AAut0I6n0yzyjRcdn9yxBnfY65qrsAbQbNmLSa_457oGg33rEt3R39lt8pKF4EbJ_SFs11QOmI0I_xzI6h5YqPgy0Rn7rkpxYMwrTq8jV2SMY-Tzk8qPYbJcOLktbIUXM1YEj4lwCNIGMqHzbHf483v3cjZR1hNo1-N8yr1uBeuxkDzI9mlb6qKord_lgiXSnzf7_diEzjk_4TPfa3Q8pVn8U_yMKJK_YBEQpW8A4r88yGeD0ur2YpvUtwxeRk0t8G8IRuKg9FzzSIDAhyllbtTaUy_40cUbXKwWUKyhydrefNnUnZ9zBPVfRWduy0cuKjUSmsvDRhjonHrQzTqXYjIpnCvsJWcRf2_aC2G6Hw6HdkfZBydEpej-Xvv0MV6T9iPYEzAjsRyOIw7atBncjwVm4Qn5-BSdSRs-fNdZqEKK56rszZwT2G6f82tYTBuV23CZEOWZF7XE6w3TPDbrYS2uJyYk33CZmHfjXNAWFDxOto-LcY-_J2w0TxCvc_pV5H-Muqw5iK6M4zPI59YQKwLEmpUEmf2hANBkURU84RRJQyxSYM8UOZhiWuANlclOXSfYT3YQQ_weLzGuB74lkgxGL7XYaaKnEL5vLybJYIAxA_L3N0HF0sCJJJT4dmx67sPFxwibc95Ax0GJOrtWyNhLpgeE53YbbZuexAi9EYCDh6J7-PKJkMvuRnJreUYhrTTVAlUtaqdZyvMR4aFGVxn8P_igXvO8mjzbPUayexvScM1ykpZOnCWRaw7SgfZasjpV6iN_s_qKq_ng__77ejZZ26j5NIFfrkE0mbl8ZoJO1ywYOC73chVpR7rbiitq4zsezdzl0aOUqXU1jmQXn_BoZVe0U3skXPde43Uk50W56A5c28z4d_YKwV3uphFXzTonkF0DTJ3kftF5NhJpeuw4shLvZlhxTVda5VBOLkPg8E0HomDU-E794X8Lk1wpyc-VjejdzVDOuAeFV_RIxc83L1pNF81NGc6XNbuHTg3wMmrrzvpYVVkEe_sL84Y-j48cBXbuaccXVXOJy4g9cGUscPWPjDF0hvEp2vNlt9hYC4ffrV5uZ7tdYtwltRgjC3JP1cJnCx5OP1NghpJdxA5iPC1PNu3T8MnYfzj7xXKjxb_gKvG1YGld0dHUqNc2oEZzys5F--Su1oZxijUSC6kQxKOUum1Q4FfCUR0lPFcIES7d9ycJvuuR1nRWej_AtPomAKaUguSlV3VBkTYBU0eZ4nXDUGxtY0GF3tjz4K17X1pA9zkS-x8pyFW1wW3PVpR40AhHjQDvhoxFqGH3sqj6A9wacJqFgP3Kh3GPSxdwASYYxqYnsT7RBgkHHLZwbDTn4F0YFWvboSTTsxz_Kjufr7jw8SyS-DccXcffByy4m9e7MrxLEy9HBGuBsEIRWenwyMAv6BdpeZtL-8FL-stAhENSMr4oQzJkg_1It9i9J_4nwi9svG-pCwiaHy6v_dpVge8rkkBScbUTC509OGQQ3BfhONdzhGcVYRZRcDPDZRVD2CJXRgrlQPMINnM0dGwkIgUVQNaidfU8olS0ABtJiLvRiOO_oNyT91hKDRIexgOHYigr_5k5l7vBz3iTT_9DZ7PSgj7Oza49XLWBABUg6tmGod4B1PGw0y-ysQENsQkG1XYRVWPLYf7sHOw6Hytqg2SPDgjUf7RIU9fMRtffwT2zXYKG-LvRTIBopC10QeDDXlbJcftv2vYr7FrG9GYoNW0--GmhCNwM6OScjE-g1IojQkY7MdGeu3ECZYRRwYN98ONXk24MeWlBYtUBWSwiw559m2-Q9DtodE81F8Bn3vM68t3FCguonFBHQ97jW2edipCCIVhIZLc-cIZKapcDyNhiHzFeYdTEnoz-06Lr5odAxUi4IDiDIt7--QfkBXAqRqc3ihqemVNyg0rIYu6386tlDhNveZzqXQ-WiumjrnjQguRooggzLEjfBXFnZeWKH0SCUsDJm5RnpnBFAVg4ueoxVkAHafsRkW5bskrdQgZ5Ah24tWaBGjLAU4PMH914m5l0dqSwtfO1-Ew"}
+zbcfg:v1:{"envelopeVersion":1,"kdf":{"name":"scrypt","salt":"6ovANDTjL45oYCr3KZvlkw","cost":65536,"blockSize":8,"parallelization":2,"keyBytes":32},"cipher":{"name":"aes-256-gcm","nonce":"L5QmpLi64kLsZlVg","authTag":"Ah3MyhpqXbqF04JarX1JPg"},"ciphertext":"gNZxzcuvV5QPRiEj6Ct7r-OIMqjIBYkblWYFU_Wz17Dnq5C-n9Ks8aq2r4Lk0MKSuVV3tGirbYwOPGRkB_WAHfTLBOq1oqMEIIHhow9kj8KVRkQSEJGzzIHyt9VRSHfBk84kf3JBQn4m3dLqdHjQE_NpWP7Yr4YLebqQ3epgYitbF4e2iUBuceSoPa7GT3b8ZIxXmbEfA_I2pP9xuOKBlaRLDV4JQRd0YPZDTMM0_Fu7EA5izU3-to3D2UQxCYqilnEfeutSnkXA9KtC8An5Gjay7VsuN3PwqpGLoeMCjJDF4EvznuTUDS6SRFe374wLUl1thVFlJV1EjZub94cmtoR-GZz2JXw8H8jqm1KxtV9hJlHJa-mk0K7QZQFBi93UBq0T2Sae8yHsVo_yNnmQDi53rxFhHdRS09i1Fz3Luq4ZA-u9RGkm0keBDyNdBwKWehqEOvNLlECApo-3un57lugkjzmNK4bpmOcSoCwVOMND8YXa_4scXrCOmwm1nPZVG1GbWnNvV3vcZoRAmT7lCUNLve1fYqtmvRd0OiSK4rCA0CLKmAhlrHjOeL09jeHaS-q-OrrTHFbCOz6DzGFXdAip2OGR2W6eCYgduBxPTalphQSJAFrUQMGZ6Sz57D2yi4Nk8GKiCXgO66PMkKHmORXq3_omd53WfjK08IWt3QDHubIcgKi-ZOeolvdcSi9Up_20ZwALmuSnovQ8UlzB0oQWYjRKpU7FRIxVvdu7lkFthZlS2Jp2ed7xG4IzXOE6dkJIZ3jzpgtl72W6WPGMXCYCcW2A3RRJ0VVKnQyEQ-NyUGG6TRv1aIjBmn2kJ-hFF1uyw1lVkwGs-9M_7k_ilLwgF6EX9pRFJTwsLe9hC_CUVaXbpp_pc5y15AdLiSQmYtAClRY2e5g0DQvksBGFGy80m65QrjPkdZkVHtAsGcyjIASw5GIa_hff-PItlsNXLldUQYYC5dVw-8XuuZKxmI9hqrxA2rDNsHfOlCkKHxIWWHQR_pkVFEM4OUAPvyZ_QU52iMAnXf3dRuyOMU_2v1duu6qbZl-MUTP2V3VGPKkbI7M0Y12D60EzExobsHO8xGpx2QVJczqHtjV0KBWFG4m355Y4-zGJUzD-y3ghW1pTKMH28rxl88w9zEnxCAADKTH8mCPFROQOF_70kHgNZhgHmS6pUyDe6dmtUEpTAM7tQSJ3Fxb7YtaDog6NWGALmBdqUliFH3_gcrR-cjUwjoj_vfjFKNiFYXhEzHaksG3MQF6ui7IMHme1sB1r0gCvT7CChbVaxpibD-MeaY9pxK00ogy10NjB2VRsH9O-ngS-X4ven-vjT9WA3m1LzPpeFH21Boa0Qvug_thbxuzeFBg1m7_xYlLChb7VnYi3lBxmtEUOukojbkoFlvV_cVztEo0CExJyQyCaVkFM2ZathZkbr1H_LdHvEySihzzEhFJA0L2R6j5YUU_CkjttV7QkYprlApH_-pccNeBN7QnUMBwtDcQt_rImBreCzRGKQ0Nu-HKmJZar9DzRI4JPdk8c3x3UCwQNJtBYgdvt4p1Vc_fUdDvI5HsR_cUwAs7uGY_gwqhBXnHqWbHKY8kAvVKTHh2Q07XCpS4gnkEFNQkgr7rN2UlSVfJ3cP6TukOc-QAZN5yjGFfWU3MMUYPUoCxeQQcZPEY84Kr9KBco7Kpy3W9O9hnAzdm5A1UkZZ0SHA_eZTfMx2R9NfrZ3nUVHwCZuyN6afmbb5AzHGw2e91TjLP8-oSSNJoU5MKULM6SKn-rMnG3WG1d6NNm5c4JsmqbGs8CO2NwnY8mQ6TBZrMaDFaz2SuILKUOchbMj5G3gJV5j47k5rQSCA_rYezvPOxNQu07zMNNaQS3vl2QATG5lH3JiOmbMLRH0EJr7y7FcYxvFA09POM_7OFw2Kg2zV0xDaHmV86qrQqUmzg2URalt6aGWAlR2Az-NhlSciaZNbiMfAnplxhr-7YsBYN2hwVpZqvOCQvZuE6AoSQDsHgGpf0__t0oqeldXJGJezTYQgTL8oiE9gSgPyUCrFVcw4e0_W3Fe-Ab2IchQ--GdWVCkAJCGLtvpvS9ugXeJZCGbv_xlZGYr429lL2LbAcVy3cBM8_88C35bPAeSYPeL4xGVe_Pl0pz2nCKAFnA32m3bkrhvkQq1Df46kos2y5Zc7Rerhixu9_zejUKsraspg7zlNSNdkdOC2OaYpi4uqsrM2X5_WlD-L0DRl7nYhnlKGJLelArb-m7tBZLR4Jl5dPWgL7_fYS5URGA6qFDSZRwMeliDRPCDEMZO1hwenCWEa6mCmXfCMRB09Ajp6OiA6TgbT2Ru6n0bOAWAAbsFQLDu5KFikjAGSCipl8yWbk4KV0BbjzPrMuoHwvBpsNl8-VTtWVMND3X5u7y-AGo8VPfhXimV7aTzmk-I3NSWyJ6n5zAdmM0TOWtJyfJhz1Fa6TepAFdSVVEIxIneTGcxhXd2dEPqvQK7FZyGijHJM79lTf2yDbJaPn5TlkuimHCNjmyfvNhplogm6enyxvufCCXWes5oD-1Gz3rOj3woqav-PQJvYSHlgJ20T5h9mbPVX0EGMyHOkLs3-IKdu8Is2ll0zwTWyLSAuRPScfGOnVgoxwbeSOTeVfKi_qLEdq_A4GpFjvz9VG6xq3c-qGSQPieoq5UWONLR8TmlmB1pqIdWB1JRGKWpsHB2xVTMlfpc5TMX3slMTVmXwv59paNVwt_L2GdJRA743s8yzZpXzEbuG6AZ0PJamTeqmL6uqjH7FIyy8Kpn4JLI0emfmdbnLj7rP8KqP0kJEx9pEZ05ePwa2rhkVGIXCAMJwHaUWfKEniWll9LMh430rkw7vj33sDYOavujH1mf9x7ZDLlXfEy9Yv9gs_FTUf197hDXr3DSO4LRk-T4MzcIn9bgtUhsDJqfD0PLp5ZFsQxKJClNBfa6X5GHT_8scFjZLDhEwf2tUehE1EGDsqAvsTvrU09NLha88iKNmKB-MXmRkp-s-uoktL7nWfE0tIA35X1EA-lQSta6b9o5TVGmpZTiXB9PS4CEvkgiT4ms1lbnTSYXeokrlw4ofAiybsY5qYlDQ7a5LB2ePfKtcDjyf2NAZMFJBxyAHqw_4lcKoOSgh0iaqqa9RqQKX1jd9fcm3L0hlAC1lGG7qhb0f1qJSKiEMncR0jO9hQpolP9l5qQMQoaghIMjDfuZ-OdYxWU-FwQ7rPayL5YJ0DGl4ycQS6NTJFvvxg_FZO3TmfGw7UxH_0Pl5kWHgK14RUM_9OeQ3O1ighjNeOsEIW-rIwajo_7eMXnoEFyKWDJzFXZxdxEMJkArlV3eFk1Jd2V4Mr1RgMN2eltuDK0Fy0W93R74aYI4ZPD2Ifg0kWDGtGFUSqBU7vWOzfBK2eRhsMGbwa_5wDqnGr5vskQnZtuzxh9P0yCePa0aIZ9Y2fZLJ0"}
