+++ app/server/modules/system/config-transfer/index.ts
@@ -1,2 +1,2 @@
 export { createPassphraseProtectedOrganizationConfigExport, OrganizationResticPasswordNotFoundError } from "./export";
-export { importPassphraseProtectedOrganizationConfig } from "./import";
+export { importConfig } from "./import";
