+++ app/server/modules/system/config-transfer/prepare-import.ts
@@ -1,37 +1,19 @@
 import { validateCustomResticParams } from "@zerobyte/core/restic/server";
 import { config as serverConfig } from "~/server/core/config";
-import { isValidCron } from "~/server/modules/backups/backup.helpers";
+import { validateScheduleTiming } from "~/server/modules/backups/backup.helpers";
 import { isValidBackupScheduleName } from "~/server/modules/backups/backup-schedule-name";
-import { mapNotificationConfigSecrets } from "~/server/modules/notifications/notification-config-secrets";
+import { encryptNotificationConfig } from "~/server/modules/notifications/notification-config-secrets";
 import { assertNotificationTargetAllowed } from "~/server/modules/notifications/utils/notification-target-policy";
-import { mapRepositoryConfigSecrets } from "~/server/modules/repositories/repository-config-secrets";
-import { mapVolumeConfigSecrets } from "~/server/modules/volumes/volume-config-secrets";
+import { encryptRepositoryConfig } from "~/server/modules/repositories/repository-config-secrets";
+import { encryptVolumeConfig } from "~/server/modules/volumes/volume-config-secrets";
 import { checkMirrorCompatibility } from "~/server/utils/backend-compatibility";
 import { cryptoUtils } from "~/server/utils/crypto";
 import { normalizeRequiredName } from "~/server/utils/names";
 import { InvalidConfigTransferError } from "./errors";
 import type { ConfigTransferModel } from "./model";
 
-export type PreparedConfigImport = Omit<ConfigTransferModel, "resticPassword"> & {
+export type PreparedImport = Omit<ConfigTransferModel, "resticPassword"> & {
 	sealedResticPassword: string;
-};
-
-const pushUnique = (items: string[], value: string) => {
-	if (!items.includes(value)) {
-		items.push(value);
-	}
-};
-
-const joinWithAnd = (items: string[]) => {
-	if (items.length <= 1) {
-		return items[0] ?? "";
-	}
-
-	if (items.length === 2) {
-		return `${items[0]} and ${items[1]}`;
-	}
-
-	return `${items.slice(0, -1).join(", ")}, and ${items[items.length - 1]}`;
 };
 
 const validateSchedules = (schedules: ConfigTransferModel["backupSchedules"]) => {
@@ -43,11 +25,8 @@ const validateSchedules = (schedules: ConfigTransferModel["backupSchedules"]) =>
 		}
 		scheduleNames.add(schedule.name);
 
-		if (schedule.cronExpression && !isValidCron(schedule.cronExpression)) {
-			throw new InvalidConfigTransferError();
-		}
-
-		if (schedule.enabled && !schedule.cronExpression) {
+		const error = validateScheduleTiming(schedule);
+		if (error) {
 			throw new InvalidConfigTransferError();
 		}
 
@@ -60,7 +39,7 @@ const validateSchedules = (schedules: ConfigTransferModel["backupSchedules"]) =>
 	}
 };
 
-const normalizeImportedResourceNames = <Resource extends { name: string }>(resources: Resource[]) => {
+const normalizeNames = <Resource extends { name: string }>(resources: Resource[]) => {
 	return resources.map((resource) => {
 		const name = normalizeRequiredName(resource.name);
 		if (name === null) {
@@ -72,21 +51,16 @@ const normalizeImportedResourceNames = <Resource extends { name: string }>(resou
 };
 
 const validateUniqueVolumeNames = (volumes: ConfigTransferModel["volumes"]) => {
-	const volumeNames = new Set<string>();
-
-	for (const volume of volumes) {
-		if (volumeNames.has(volume.name)) {
-			throw new InvalidConfigTransferError();
-		}
-
-		volumeNames.add(volume.name);
+	const volumeNames = new Set(volumes.map((volume) => volume.name));
+	if (volumeNames.size !== volumes.length) {
+		throw new InvalidConfigTransferError();
 	}
 };
 
 const normalizeImportedNames = (payload: ConfigTransferModel): ConfigTransferModel => {
-	const repositories = normalizeImportedResourceNames(payload.repositories);
-	const volumes = normalizeImportedResourceNames(payload.volumes);
-	const notificationDestinations = normalizeImportedResourceNames(payload.notificationDestinations);
+	const repositories = normalizeNames(payload.repositories);
+	const volumes = normalizeNames(payload.volumes);
+	const notificationDestinations = normalizeNames(payload.notificationDestinations);
 	validateUniqueVolumeNames(volumes);
 
 	return { ...payload, repositories, volumes, notificationDestinations };
@@ -129,7 +103,7 @@ const createLocalPathWarnings = (payload: ConfigTransferModel) => {
 	const warnings: string[] = [];
 	const volumeRequirements = new Map<string, string>();
 	const repositoryRequirements = new Map<string, string>();
-	const mirrorRequirementsBySchedule = new Map<string, string[]>();
+	const mirrorRequirements = new Map<string, Set<string>>();
 
 	for (const volume of payload.volumes) {
 		if (volume.config.backend === "directory") {
@@ -155,44 +129,45 @@ const createLocalPathWarnings = (payload: ConfigTransferModel) => {
 			continue;
 		}
 
-		const requirements = mirrorRequirementsBySchedule.get(mirror.scheduleRef) ?? [];
-		pushUnique(requirements, requirement);
-		mirrorRequirementsBySchedule.set(mirror.scheduleRef, requirements);
+		const requirements = mirrorRequirements.get(mirror.scheduleRef) ?? new Set<string>();
+		requirements.add(requirement);
+		mirrorRequirements.set(mirror.scheduleRef, requirements);
 	}
 
-	const requirementsBySchedule = new Map<string, string[]>();
+	const schedulesRequiringReview = new Set<string>();
 
 	for (const schedule of payload.backupSchedules) {
-		const requirements: string[] = [];
+		const requirements = new Set<string>();
 		const volumeRequirement = volumeRequirements.get(schedule.volumeRef);
 		const repositoryRequirement = repositoryRequirements.get(schedule.repositoryRef);
 
 		if (volumeRequirement) {
-			requirements.push(volumeRequirement);
+			requirements.add(volumeRequirement);
 		}
 		if (repositoryRequirement) {
-			requirements.push(repositoryRequirement);
+			requirements.add(repositoryRequirement);
 		}
-		for (const mirrorRequirement of mirrorRequirementsBySchedule.get(schedule.ref) ?? []) {
-			pushUnique(requirements, mirrorRequirement);
+		for (const mirrorRequirement of mirrorRequirements.get(schedule.ref) ?? []) {
+			requirements.add(mirrorRequirement);
 		}
 
-		if (requirements.length > 0) {
-			requirementsBySchedule.set(schedule.ref, requirements);
+		if (requirements.size > 0) {
+			schedulesRequiringReview.add(schedule.ref);
 			if (schedule.enabled) {
+				const requirementList = new Intl.ListFormat("en", { type: "conjunction" }).format(requirements);
 				warnings.push(
-					`Disabled schedule "${schedule.name}" because it references ${joinWithAnd(requirements)}. Re-enable it after validating those imported paths on this server.`,
+					`Disabled schedule "${schedule.name}" because it references ${requirementList}. Re-enable it after validating those imported paths on this server.`,
 				);
 			}
 		}
 	}
 
-	return { requirementsBySchedule, warnings };
+	return { schedulesRequiringReview, warnings };
 };
 
-export const prepareConfigImport = async (
+export const prepareImport = async (
 	payload: ConfigTransferModel,
-): Promise<{ prepared: PreparedConfigImport; warnings: string[] }> => {
+): Promise<{ prepared: PreparedImport; warnings: string[] }> => {
 	const normalizedPayload = normalizeImportedNames(payload);
 	validateSchedules(normalizedPayload.backupSchedules);
 	await validateMirrors(normalizedPayload);
@@ -205,48 +180,42 @@ export const prepareConfigImport = async (
 		throw new InvalidConfigTransferError();
 	}
 
-	const sealImportedSecret = async (value: string) => {
-		try {
-			return await cryptoUtils.sealSecret(value);
-		} catch {
-			throw new InvalidConfigTransferError();
-		}
-	};
-	const sealImportedResticPassword = async () => {
-		try {
-			return await cryptoUtils.sealSecret(normalizedPayload.resticPassword);
-		} catch {
-			throw new InvalidConfigTransferError();
-		}
-	};
+	let repositories: PreparedImport["repositories"];
+	let volumes: PreparedImport["volumes"];
+	let notificationDestinations: PreparedImport["notificationDestinations"];
+	let sealedResticPassword: string;
+	try {
+		[repositories, volumes, notificationDestinations, sealedResticPassword] = await Promise.all([
+			Promise.all(
+				normalizedPayload.repositories.map(async (repository) => ({
+					...repository,
+					config: await encryptRepositoryConfig(repository.config),
+				})),
+			),
+			Promise.all(
+				normalizedPayload.volumes.map(async (volume) => ({
+					...volume,
+					config: await encryptVolumeConfig(volume.config),
+				})),
+			),
+			Promise.all(
+				normalizedPayload.notificationDestinations.map(async (destination) => ({
+					...destination,
+					config: await encryptNotificationConfig(destination.config),
+				})),
+			),
+			cryptoUtils.sealSecret(normalizedPayload.resticPassword),
+		]);
+	} catch {
+		throw new InvalidConfigTransferError();
+	}
 
-	const [repositories, volumes, notificationDestinations, sealedResticPassword] = await Promise.all([
-		Promise.all(
-			normalizedPayload.repositories.map(async (repository) => ({
-				...repository,
-				config: await mapRepositoryConfigSecrets(repository.config, sealImportedSecret),
-			})),
-		),
-		Promise.all(
-			normalizedPayload.volumes.map(async (volume) => ({
-				...volume,
-				config: await mapVolumeConfigSecrets(volume.config, sealImportedSecret),
-			})),
-		),
-		Promise.all(
-			normalizedPayload.notificationDestinations.map(async (destination) => ({
-				...destination,
-				config: await mapNotificationConfigSecrets(destination.config, sealImportedSecret),
-			})),
-		),
-		sealImportedResticPassword(),
-	]);
+	const { schedulesRequiringReview, warnings } = createLocalPathWarnings(normalizedPayload);
+	const backupSchedules = normalizedPayload.backupSchedules.map((schedule) => {
+		const enabled = schedule.enabled && !schedulesRequiringReview.has(schedule.ref);
 
-	const { requirementsBySchedule, warnings } = createLocalPathWarnings(normalizedPayload);
-	const backupSchedules = normalizedPayload.backupSchedules.map((schedule) => ({
-		...schedule,
-		enabled: requirementsBySchedule.has(schedule.ref) ? false : schedule.enabled,
-	}));
+		return { ...schedule, enabled };
+	});
 	const { resticPassword: _resticPassword, ...config } = normalizedPayload;
 
 	return {
